Some cleanup on file loading
